Move `disabled_on` to its own model
