+++ datahub/core/models.py
@@ -56,6 +56,21 @@ class ArchivableModel(models.Model):
         self.save()
 
 
+class DisableableModel(models.Model):
+    """Defines a disabled_on field and related logic."""
+
+    disabled_on = models.DateTimeField(blank=True, null=True)
+
+    def was_disabled_on(self, date_on):
+        """Returns True if this object was disabled at time `date_on`, False otherwise."""
+        if not self.disabled_on:
+            return False
+        return self.disabled_on <= date_on
+
+    class Meta:  # noqa: D101
+        abstract = True
+
+
 class BaseConstantModel(models.Model):
     """Constant tables for FKs."""
